filter: enable percentage based rollouts
Currently, we only support checking if a feature is fully rolled out.
This change gives us percentage based rollouts by getting the last two
digits from the full computed hash of the flag, and checking if it's
less than the rollout percentage.

This is fine for now, but I need to check if the distribution for those
two digits is consistent enough to actually use this forever. It may be
that the distribution is uneven.